+++ src/ViewInstance.js
@@ -20,7 +20,7 @@ import Headline from '@folio/stripes-components/lib/Headline';
 import MultiColumnList from '@folio/stripes-components/lib/MultiColumnList';
 import ViewMetaData from '@folio/stripes-smart-components/lib/ViewMetaData';
 
-import craftLayerUrl from '@folio/stripes-components/util/craftLayerUrl';
+import { craftLayerUrl } from './utils';
 
 import formatters from './referenceFormatters';
 
@@ -206,7 +206,7 @@ class ViewInstance extends React.Component {
         <IconButton
           id="clickable-edit-instance"
           style={{ visibility: !instance ? 'hidden' : 'visible' }}
-          href={this.craftLayerUrl('edit')}
+          href={this.craftLayerUrl('edit', location)}
           onClick={this.onClickEditInstance}
           title={formatMsg({ id: 'ui-inventory.editInstance' })}
           icon="edit"
@@ -232,7 +232,7 @@ class ViewInstance extends React.Component {
     const newHoldingsRecordButton = (
       <Button
         id="clickable-new-holdings-record"
-        href={this.craftLayerUrl('createHoldingsRecord')}
+        href={this.craftLayerUrl('createHoldingsRecord', location)}
         onClick={this.onClickAddNewHoldingsRecord}
         title={formatMsg({ id: 'ui-inventory.addHoldings' })}
         buttonStyle="primary"
@@ -426,9 +426,9 @@ class ViewInstance extends React.Component {
             </Col>
           </Row>
           <Row>
-            { (instance.instanceFormatId) &&
-              <Col xs={3}>
-                <KeyValue label={formatMsg({ id: 'ui-inventory.format' })} value={formatters.instanceFormatsFormatter(instance, referenceTables.instanceFormats)} />
+            { (instance.instanceFormatIds && instance.instanceFormatIds.length > 0) &&
+              <Col xs={6}>
+                <KeyValue label={formatMsg({ id: 'ui-inventory.format' })} value={_.get(instance, ['instanceFormatIds'], []).map((formatId, i) => <div key={i}>{this.refLookup(referenceTables.instanceFormats, formatId).name}</div>)} />
               </Col>
             }
           </Row>