+++ src/edit/InstanceForm.js
@@ -29,6 +29,7 @@ import PublicationFrequencyFields from './publicationFrequencyFields';
 import PublicationRangeFields from './publicationRangeFields';
 import NoteFields from './noteFields';
 import ElectronicAccessFields from './electronicAccessFields';
+import InstanceFormatFields from './instanceFormatFields';
 import LanguageFields from './languageFields';
 import ChildInstanceFields from './childInstanceFields';
 import ParentInstanceFields from './parentInstanceFields';
@@ -197,15 +198,6 @@ class InstanceForm extends React.Component {
     const instanceStatusOptions = referenceTables.instanceStatuses ? referenceTables.instanceStatuses.map(
       it => ({
         label: `${it.name} (${it.source}: ${it.code})`,
-        value: it.id,
-        selected: it.id === initialValues.instanceFormatId,
-      }),
-    ) : [];
-
-
-    const instanceFormatOptions = referenceTables.instanceFormats ? referenceTables.instanceFormats.map(
-      it => ({
-        label: it.name,
         value: it.id,
         selected: it.id === initialValues.instanceFormatId,
       }),
@@ -381,13 +373,7 @@ class InstanceForm extends React.Component {
                   dataOptions={[{ label: formatMsg({ id: 'ui-inventory.selectResourceType' }), value: '' }, ...instanceTypeOptions]}
                   required
                 />
-                <Field
-                  name="instanceFormatId"
-                  type="text"
-                  component={Select}
-                  label={formatMsg({ id: 'ui-inventory.format' })}
-                  dataOptions={[{ label: formatMsg({ id: 'ui-inventory.selectFormat' }), value: '' }, ...instanceFormatOptions]}
-                />
+                <InstanceFormatFields instanceFormats={referenceTables.instanceFormats} formatMsg={formatMsg} />
                 <LanguageFields formatMsg={formatMsg} />
                 <PublicationFrequencyFields formatMsg={formatMsg} />
                 <PublicationRangeFields formatMsg={formatMsg} />